+++ CinemaCat.Application.Tests/Movies/SearchMovieHandlerTests.cs
@@ -1,24 +1,24 @@
-﻿using CinemaCat.Application.Handlers.Persons.SearchPerson;
+﻿using CinemaCat.Application.Handlers.Movies.SearchMovie;
 using CinemaCat.Application.Interfaces;
 using CinemaCat.Domain.Models;
 using System.Linq.Expressions;
 
-namespace CinemaCat.Application.Tests;
+namespace CinemaCat.Application.Tests.Movies;
 
-public class SearchPersonHandlerTests
+public class SearchMovieHandlerTests
 {
     private readonly IFixture _fixture = new Fixture().Customize(new DateOnlyFixtureCustomization());
-    private readonly Mock<IDataBaseProvider<PersonDetails>> _dataBaseProviderMock = new();
+    private readonly Mock<IDataBaseProvider<Movie>> _dataBaseProviderMock = new();
 
     [Fact]
     public async Task Handle_ThrowsException_Error()
     {
         // arrange
-        var request = _fixture.Create<SearchPersonPequest>();
+        var request = _fixture.Create<SearchMovieRequest>();
         var exception = new Exception("a weird error");
-        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<PersonDetails, bool>>>()))
+        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<Movie, bool>>>()))
             .ThrowsAsync(exception);
-        var _handler = new SearchPersonHandler(_dataBaseProviderMock.Object);
+        var _handler = new SearchMovieHandler(_dataBaseProviderMock.Object);
 
         // act
         var response = await _handler.Handle(request, CancellationToken.None);
@@ -33,11 +33,11 @@ public class SearchPersonHandlerTests
     public async Task Handle_Success_ReturnsResult()
     {
         // arrange
-        var request = _fixture.Create<SearchPersonPequest>();
-        var man = _fixture.Create<PersonDetails>() with { Name = request.Name };
-        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<PersonDetails, bool>>>()))
-            .ReturnsAsync(new List<PersonDetails> { man });
-        var _handler = new SearchPersonHandler(_dataBaseProviderMock.Object);
+        var request = _fixture.Create<SearchMovieRequest>();
+        var movie = _fixture.Create<Movie>() with { Title = request.Title };
+        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<Movie, bool>>>()))
+            .ReturnsAsync(new List<Movie> { movie });
+        var _handler = new SearchMovieHandler(_dataBaseProviderMock.Object);
 
         // act
         var response = await _handler.Handle(request, CancellationToken.None);
@@ -45,6 +45,6 @@ public class SearchPersonHandlerTests
         // assert
         response.IsSuccess.Should().BeTrue();
         response.Result.Should().NotBeNull();
-        response.Result.First().Should().Be(man);
+        response.Result.First().Should().Be(movie);
     }
 }
