+++ CinemaCat.Application.Tests/Persons/SearchPersonHandlerTests.cs
@@ -1,10 +1,11 @@
-﻿using CinemaCat.Application.Handlers.Persons.GetPerson;
+﻿using CinemaCat.Application.Handlers.Persons.SearchPerson;
 using CinemaCat.Application.Interfaces;
 using CinemaCat.Domain.Models;
+using System.Linq.Expressions;
 
-namespace CinemaCat.Application.Tests;
+namespace CinemaCat.Application.Tests.Persons;
 
-public class GetPersonHandlerTests
+public class SearchPersonHandlerTests
 {
     private readonly IFixture _fixture = new Fixture().Customize(new DateOnlyFixtureCustomization());
     private readonly Mock<IDataBaseProvider<PersonDetails>> _dataBaseProviderMock = new();
@@ -13,11 +14,11 @@ public class GetPersonHandlerTests
     public async Task Handle_ThrowsException_Error()
     {
         // arrange
-        var request = _fixture.Create<GetPersonRequest>();
+        var request = _fixture.Create<SearchPersonPequest>();
         var exception = new Exception("a weird error");
-        _dataBaseProviderMock.Setup(m => m.GetByIdAsync(It.IsAny<Guid>()))
+        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<PersonDetails, bool>>>()))
             .ThrowsAsync(exception);
-        var _handler = new GetPersonHandler(_dataBaseProviderMock.Object);
+        var _handler = new SearchPersonHandler(_dataBaseProviderMock.Object);
 
         // act
         var response = await _handler.Handle(request, CancellationToken.None);
@@ -32,11 +33,11 @@ public class GetPersonHandlerTests
     public async Task Handle_Success_ReturnsResult()
     {
         // arrange
-        var request = _fixture.Create<GetPersonRequest>();
-        var person = _fixture.Create<PersonDetails>() with { Id = request.Id };
-        _dataBaseProviderMock.Setup(m => m.GetByIdAsync(person.Id))
-            .ReturnsAsync(person);
-        var _handler = new GetPersonHandler(_dataBaseProviderMock.Object);
+        var request = _fixture.Create<SearchPersonPequest>();
+        var man = _fixture.Create<PersonDetails>() with { Name = request.Name };
+        _dataBaseProviderMock.Setup(p => p.GetAsync(It.IsAny<Expression<Func<PersonDetails, bool>>>()))
+            .ReturnsAsync(new List<PersonDetails> { man });
+        var _handler = new SearchPersonHandler(_dataBaseProviderMock.Object);
 
         // act
         var response = await _handler.Handle(request, CancellationToken.None);
@@ -44,6 +45,6 @@ public class GetPersonHandlerTests
         // assert
         response.IsSuccess.Should().BeTrue();
         response.Result.Should().NotBeNull();
-        response.Result.Should().Be(person);
+        response.Result.First().Should().Be(man);
     }
 }
