+++ CinemaCat.Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -23,6 +23,8 @@ public static class ServiceCollectionExtensions
     {
         ConfigureBsonTypesMapping();
 
+        serviceCollection.Configure<MongoConfiguration>(configuration.GetSection(MongoConfiguration.SectionName));
+
         serviceCollection.AddSingleton<MongoClient>(provider =>
         {
             var options = provider.GetRequiredService<IOptions<MongoConfiguration>>();
